// manager_frontend/src/client/models/ManagerCustomerCreatePayload.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ManagerCustomerCreatePayload = {
    name: string;
    phone?: (string | null);
    email?: (string | null);
    type?: string;
    inn?: (string | null);
    kpp?: (string | null);
    full_legal_name?: (string | null);
    legal_address?: (string | null);
    actual_address?: (string | null);
    city?: (string | null);
    bank_name?: (string | null);
    bic?: (string | null);
    iban?: (string | null);
    signer_position?: (string | null);
    signer_name?: (string | null);
    acting_basis?: (string | null);
    signing_mode?: (string | null);
    is_favorite?: (boolean | null);
};